separating mapi.py into action, billing, webservice
